rename MorphicCore to Morphic.Core
