+++ Morphic.Client/Morphic.Client.csproj
@@ -50,7 +50,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MorphicCore\MorphicCore.csproj" />
+    <ProjectReference Include="..\Morphic.Core\Morphic.Core.csproj" />
     <ProjectReference Include="..\MorphicService\MorphicService.csproj" />
     <ProjectReference Include="..\MorphicSettings\MorphicSettings.csproj" />
   </ItemGroup>
