+++ Morphic.Client/UserSettings.cs
@@ -24,7 +24,7 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
-using MorphicCore;
+using Morphic.Core;
 
 namespace Morphic.Client
 {
