+++ Morphic.Core/IRecord.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace MorphicCore
+namespace Morphic.Core
 {
     /// <summary>
     /// A Storable record with a unique identifier
