+++ Morphic.Core/Morphic.Core.csproj
