+++ MorphicService/MorphicService.csproj
@@ -11,7 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MorphicCore\MorphicCore.csproj" />
+    <ProjectReference Include="..\Morphic.Core\Morphic.Core.csproj" />
     <ProjectReference Include="..\MorphicSettings\MorphicSettings.csproj" />
   </ItemGroup>
