+++ MorphicSettings.Tests/MorphicSettings.Tests.csproj
@@ -14,7 +14,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MorphicCore\MorphicCore.csproj" />
     <ProjectReference Include="..\MorphicSettings\MorphicSettings.csproj" />
   </ItemGroup>
