+++ MorphicSettings/CaptureSession.cs
@@ -25,7 +25,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using System.Text;
-using MorphicCore;
+using Morphic.Core;
 
 namespace MorphicSettings
 {
