+++ MorphicSettings/SettingHandler.cs
@@ -23,7 +23,7 @@
 
 using System;
 using System.Threading.Tasks;
-using MorphicCore;
+using Morphic.Core;
 using System.Collections.Generic;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
