+++ MorphicSettings/Settings.cs
@@ -27,7 +27,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using MorphicCore;
+using Morphic.Core;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.DependencyInjection;
