+++ MorphicWin.sln
@@ -3,21 +3,21 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29806.167
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MorphicCore", "MorphicCore\MorphicCore.csproj", "{31FB0877-17D1-444F-9506-EE5887A63AF6}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MorphicService", "MorphicService\MorphicService.csproj", "{C3C85B67-282B-4D19-AF64-0EE027A9D307}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MorphicSettings", "MorphicSettings\MorphicSettings.csproj", "{2AD5A5C8-5538-48C8-838D-97F42ED4ECE1}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Morphic.Windows.Native", "Morphic.Windows.Native\Morphic.Windows.Native.csproj", "{98AC1F49-CE2F-4294-8FAD-1A287B0E7824}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MorphicCore.Tests", "MorphicCore.Tests\MorphicCore.Tests.csproj", "{66231E1C-DED9-4BE3-A6D2-99A65ACC379A}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MorphicSettings.Tests", "MorphicSettings.Tests\MorphicSettings.Tests.csproj", "{BE41A067-66B6-4F13-AABA-910D4B8F567A}"
 EndProject
 Project("{930C7802-8A8C-48F9-8165-68863BCCD9DD}") = "Morphic.Setup", "Morphic.Setup\Morphic.Setup.wixproj", "{8B686C6D-D3DA-4019-B4B7-A0643F17A7F7}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Morphic.Client", "Morphic.Client\Morphic.Client.csproj", "{2485B22B-5C46-46F1-A466-F8A5B7E28C64}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Morphic.Core", "Morphic.Core\Morphic.Core.csproj", "{F05E2E47-8F9B-4C41-827A-2BBC23C1B276}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Morphic.Core.Tests", "Morphic.Core.Tests\Morphic.Core.Tests.csproj", "{2FECA817-61F1-4E33-8CDC-2918B87CA5EF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,14 +27,6 @@ Global
 		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{31FB0877-17D1-444F-9506-EE5887A63AF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{31FB0877-17D1-444F-9506-EE5887A63AF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{31FB0877-17D1-444F-9506-EE5887A63AF6}.Debug|x64.ActiveCfg = Debug|x64
-		{31FB0877-17D1-444F-9506-EE5887A63AF6}.Debug|x64.Build.0 = Debug|x64
-		{31FB0877-17D1-444F-9506-EE5887A63AF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{31FB0877-17D1-444F-9506-EE5887A63AF6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{31FB0877-17D1-444F-9506-EE5887A63AF6}.Release|x64.ActiveCfg = Release|x64
-		{31FB0877-17D1-444F-9506-EE5887A63AF6}.Release|x64.Build.0 = Release|x64
 		{C3C85B67-282B-4D19-AF64-0EE027A9D307}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C3C85B67-282B-4D19-AF64-0EE027A9D307}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C3C85B67-282B-4D19-AF64-0EE027A9D307}.Debug|x64.ActiveCfg = Debug|x64
@@ -59,14 +51,6 @@ Global
 		{98AC1F49-CE2F-4294-8FAD-1A287B0E7824}.Release|Any CPU.Build.0 = Release|Any CPU
 		{98AC1F49-CE2F-4294-8FAD-1A287B0E7824}.Release|x64.ActiveCfg = Release|x64
 		{98AC1F49-CE2F-4294-8FAD-1A287B0E7824}.Release|x64.Build.0 = Release|x64
-		{66231E1C-DED9-4BE3-A6D2-99A65ACC379A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{66231E1C-DED9-4BE3-A6D2-99A65ACC379A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{66231E1C-DED9-4BE3-A6D2-99A65ACC379A}.Debug|x64.ActiveCfg = Debug|x64
-		{66231E1C-DED9-4BE3-A6D2-99A65ACC379A}.Debug|x64.Build.0 = Debug|x64
-		{66231E1C-DED9-4BE3-A6D2-99A65ACC379A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{66231E1C-DED9-4BE3-A6D2-99A65ACC379A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{66231E1C-DED9-4BE3-A6D2-99A65ACC379A}.Release|x64.ActiveCfg = Release|Any CPU
-		{66231E1C-DED9-4BE3-A6D2-99A65ACC379A}.Release|x64.Build.0 = Release|Any CPU
 		{BE41A067-66B6-4F13-AABA-910D4B8F567A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{BE41A067-66B6-4F13-AABA-910D4B8F567A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BE41A067-66B6-4F13-AABA-910D4B8F567A}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -87,6 +71,22 @@ Global
 		{2485B22B-5C46-46F1-A466-F8A5B7E28C64}.Release|Any CPU.Build.0 = Release|Any CPU
 		{2485B22B-5C46-46F1-A466-F8A5B7E28C64}.Release|x64.ActiveCfg = Release|x64
 		{2485B22B-5C46-46F1-A466-F8A5B7E28C64}.Release|x64.Build.0 = Release|x64
+		{F05E2E47-8F9B-4C41-827A-2BBC23C1B276}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F05E2E47-8F9B-4C41-827A-2BBC23C1B276}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F05E2E47-8F9B-4C41-827A-2BBC23C1B276}.Debug|x64.ActiveCfg = Debug|x64
+		{F05E2E47-8F9B-4C41-827A-2BBC23C1B276}.Debug|x64.Build.0 = Debug|x64
+		{F05E2E47-8F9B-4C41-827A-2BBC23C1B276}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F05E2E47-8F9B-4C41-827A-2BBC23C1B276}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F05E2E47-8F9B-4C41-827A-2BBC23C1B276}.Release|x64.ActiveCfg = Release|x64
+		{F05E2E47-8F9B-4C41-827A-2BBC23C1B276}.Release|x64.Build.0 = Release|x64
+		{2FECA817-61F1-4E33-8CDC-2918B87CA5EF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2FECA817-61F1-4E33-8CDC-2918B87CA5EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2FECA817-61F1-4E33-8CDC-2918B87CA5EF}.Debug|x64.ActiveCfg = Debug|x64
+		{2FECA817-61F1-4E33-8CDC-2918B87CA5EF}.Debug|x64.Build.0 = Debug|x64
+		{2FECA817-61F1-4E33-8CDC-2918B87CA5EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2FECA817-61F1-4E33-8CDC-2918B87CA5EF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2FECA817-61F1-4E33-8CDC-2918B87CA5EF}.Release|x64.ActiveCfg = Release|x64
+		{2FECA817-61F1-4E33-8CDC-2918B87CA5EF}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
